break round manager tests to sub modules
